fix: separate and change components in cart page
